Move MenuOptions to dedicated folder
It was getting a bit messy in the parent.
