+++ HarpGuru/OverlayMenu/MenuOptions/HarpKeyOption/nudgeHarpStrataByHarpKey/index.ts
@@ -1,7 +1,7 @@
 import { HarpStrata, getCovariantSet, getHarpStrata, getPitchIds, PitchIds } from 'harpstrata'
 
-import {DisplayModes} from '../../../HarpFace'
-import {getPropsForHarpStrata} from '../../../Controls/ControlHelpers'
+import {DisplayModes} from '../../../../HarpFace'
+import {getPropsForHarpStrata} from '../../../../Controls/ControlHelpers'
 
 const getNextId = (rootId: PitchIds, direction: 'UP' | 'DOWN'): PitchIds => {
   if (direction === 'UP') {
