+++ HarpGuru/OverlayMenu/MenuOptions/RootPitchOption/nudgeHarpStrataByRootPitch/index.test.ts
@@ -1,6 +1,6 @@
 import {HarpStrataProps, ApparatusIds, PozitionIds, PitchIds, ActiveIds, getHarpStrata} from 'harpstrata'
 
-import {DisplayModes} from '../../../HarpFace'
+import {DisplayModes} from '../../../../HarpFace'
 
 import { nudgeHarpStrataByRootPitch } from './index'
